requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml>=5.4.0
pydantic==2.5.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
python-dateutil==2.8.2

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette>=0.19.0

# Database
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # For async PostgreSQL

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Processing
numpy==1.25.2
pandas==2.1.3
scipy>=1.7.0
scikit-learn>=1.0.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly==5.17.0
dash==2.14.2

# Jupyter Environment
jupyter>=1.0.0
jupyterlab>=3.0.0
notebook>=6.4.0

# Geographic Data Processing
geopandas>=0.9.0
folium>=0.12.0

# Testing
pytest==7.4.3
pytest-cov>=2.12.0
pytest-asyncio==0.21.1
httpx==0.25.2

# Documentation
sphinx>=4.0.0
mkdocs>=1.2.0

# Code Quality
black>=21.6b0
flake8>=3.9.0
isort>=5.9.0
mypy>=0.910

# Data Quality and Validation
great-expectations>=0.13.0
pandas-profiling>=3.0.0

# Utils
pydantic-settings==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
slowapi==0.1.9

# Additional Dependencies
dependency-injector==4.41.0
redis==5.0.1
celery==5.3.4
psycopg2-binary==2.9.9
